+++ apps/client/src/environments/patch-notes.ts
@@ -1,23 +1,17 @@
 export const patchNotes = `### Bug Fixes
 
-* **alarms:** fixed weather transition timers sometimes acting up.
-* **db:** updated the list of legendary fish using japanese description matching.
-* **layout:** fixed some items not detected as housing vendor items.
-* **overlay:** fixed height not 100% when clock is disabled.
-* **profile:** removed veteran tag as it's not working as expected.
-* **universalis:** fixed possible bug with null cid.
+* **alarms:** fixed wrong alarm timers on last day before weather spawn at midnight.
+* **core:** fixed fishing spot positions off by -1/-1.
+* **desktop:** fixed gil transfers from retainer not taken into account.
+* **desktop:** only check for npcap if capture is set to npcap mode.
+* **fishing-reporter:** fixed tracking state not reset properly on character change.
+* **gearset:** fixed sets requiring one more piece to activate compared to ingame.
+* **leveling-equipment:** fixed wrong stat priorisation on lower level gear.
 
 
 ### Features
 
-* **currency-spending:** added quantity sold last week to have better sorting.
-* **db:** log tracking integration in fishing spot pages.
-* **desktop:** updated bundled npcap installer to 1.50.
-* **leveling-equipment:** added clickable item name to be able to copy to clipboard.
-* **list:** new button on alarm button group to open map.
-* **list:** you can now ignore requirements for a given item.
-
-
-### Performance Improvements
-
-* **list:** better loading performances with new menu system.`;
+* **db:** added new disclaimer regarding comments content.
+* **desktop:** new setting to minimize to tray bar with reduce button.
+* **list:** added snagging icon to revalant fish gathering details.
+* **log-tracker:** added tug to fish details.`;